requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
entrypoints==0.3
flake8==3.7.9
idna==2.9
mccabe==0.6.1
pathspec==0.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.13.0
regex==2020.4.4
requests==2.23.0
rope==0.16.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
